Bugfix: Fixed wrong suffix, signed vs unsigned problem. (now.tv_usec-start.tv_usec) could be negative, thus -5000 / 1000ul would result in a very larger positive number, instead of -5; thus Timer::getMilliseconds would behave erratically, randomly returning very larger milliseconds as time ellapsed. This fixes a near-infinite loop when particle FXs used, as reported by dermont in https://forums.ogre3d.org/viewtopic.php?f=2&t=93735&p=540996#p540987

+++ OgreMain/src/GLX/OgreTimer.cpp
@@ -53,7 +53,7 @@ unsigned long Timer::getMilliseconds()
 {
     struct timeval now;
     gettimeofday(&now, NULL);
-    return (now.tv_sec-start.tv_sec)*1000ul+(now.tv_usec-start.tv_usec)/1000ul;
+    return (now.tv_sec-start.tv_sec)*1000ul+(unsigned long)((now.tv_usec-start.tv_usec)/1000l);
 }
 
 //--------------------------------------------------------------------------------//
